Store the paged KV cache as FP16
Halves KV-cache memory and paged-decode attention bandwidth by storing key/value
pages as __half instead of f32.

Done in two verified phases behind a single `kv_t` type alias in kernels.cu:
- Phase 1 introduced `kv_t` (= float) plus overloaded kv_to_float / kv_from_float
  helpers across the 3 paged write kernels, copy_prefix, and 8 paged-decode read
  kernels — proven byte-identical to the f32 path.
- Phase 2 flips `kv_t` to __half and sizes the paged pools (gpu.rs) and the KV
  manager accounting (lib.rs) to 2 bytes/element.

The contiguous (non-paged) KV cache and its kernels stay f32 — only the paged
pool that feeds the paged kernels changed. nvcc type-checks every launcher cast.

Verified on Qwen2.5-Coder-3B (RTX A2000): KV page bytes halved (pool 576->288
MB), decode output and prefix reuse unchanged, agent loop 286s -> 170s. The
text/tiled-paged path is exercised; the flash/generic/MRoPE paged variants were
converted uniformly but are fallbacks not hit by this GPU/model and warrant
their own parity check.

+++ crates/hi-cuda/src/gpu.rs
@@ -15449,10 +15449,10 @@ mod native {
                 .and_then(|value| value.checked_mul(dims.v_head_dim))
                 .context("CUDA paged KV value cache element count overflows usize")?;
             let key_bytes = key_elements
-                .checked_mul(std::mem::size_of::<f32>())
+                .checked_mul(std::mem::size_of::<u16>()) // f16 KV pages (kv_t = __half)
                 .context("CUDA paged KV key cache byte count overflows usize")?;
             let value_bytes = value_elements
-                .checked_mul(std::mem::size_of::<f32>())
+                .checked_mul(std::mem::size_of::<u16>()) // f16 KV pages (kv_t = __half)
                 .context("CUDA paged KV value cache byte count overflows usize")?;
             let page_table = (0..page_table_len)
                 .map(|page| u32::try_from(page).context("CUDA KV page index does not fit u32"))
@@ -15556,10 +15556,10 @@ mod native {
                 .and_then(|value| value.checked_mul(dims.v_head_dim))
                 .context("CUDA paged batch device value pool element count overflows usize")?;
             let key_bytes = key_elements
-                .checked_mul(std::mem::size_of::<f32>())
+                .checked_mul(std::mem::size_of::<u16>()) // f16 KV pages (kv_t = __half)
                 .context("CUDA paged batch device key pool byte count overflows usize")?;
             let value_bytes = value_elements
-                .checked_mul(std::mem::size_of::<f32>())
+                .checked_mul(std::mem::size_of::<u16>()) // f16 KV pages (kv_t = __half)
                 .context("CUDA paged batch device value pool byte count overflows usize")?;
             let mut layers = Vec::with_capacity(layer_count);
             for _ in 0..layer_count {
@@ -15655,10 +15655,10 @@ mod native {
                 .and_then(|value| value.checked_mul(dims.v_head_dim))
                 .context("CUDA paged batch KV value cache element count overflows usize")?;
             let key_bytes = key_elements
-                .checked_mul(std::mem::size_of::<f32>())
+                .checked_mul(std::mem::size_of::<u16>()) // f16 KV pages (kv_t = __half)
                 .context("CUDA paged batch KV key cache byte count overflows usize")?;
             let value_bytes = value_elements
-                .checked_mul(std::mem::size_of::<f32>())
+                .checked_mul(std::mem::size_of::<u16>()) // f16 KV pages (kv_t = __half)
                 .context("CUDA paged batch KV value cache byte count overflows usize")?;
             let mut page_table = Vec::with_capacity(physical_pages);
             for batch in 0..batch_count {
